From 4df3ded5addb80690d312dfdd1483ae682bd80db Mon Sep 17 00:00:00 2001 Subject: [PATCH RFC  1/5] powerpc:numa Add numa_cpu_lookup function to update  lookup table
We access numa_cpu_lookup_table array directly in all the places
to read/update numa cpu lookup information. Instead use a helper
function to update.

This is helpful in changing the way numa<-->cpu mapping in single
place when needed.

This is a cosmetic change, no change in functionality.

Signed-off-by: Raghavendra K T <raghavendra.kt@linux.vnet.ibm.com>